refactor(gke): split version resolution logic into version.go
Pure computation (ResolveVersionForChannel and helpers) separated from
the API client in gke.go. Also demote Option 2 to non-recommended in
image-management.md — channel tracking is the single recommended path.
